added responsive header, changed overall design
now when you click on one of the header links the .active is aplied
to the clicked link and is removed from other links. The color scheme
has been changed for the worse or the better. The body has not been
changed yet and is rendered unreadable on hover. The issue will be fixed
on 08/10/22.